Dinamic Component Leccion,Hystory.. etx
