+++ frontend/src/components/pages/Mitologia_.jsx
@@ -1,10 +1,7 @@
 //librerias
 import React from 'react';
 import Header from '../Home/Header';
-import Navegator from '../profile/Navegator';
-import icon_profile from '../../assets/logo/profile_default.svg';
-import Section from '../history/Section';
-import { useNavigate } from 'react-router-dom';
+import RenderSections from '../utils/RenderSections';
 
 //Informacion de la seccion mitologia
 const DATA_HISTORY = [
@@ -23,35 +20,10 @@ const DATA_HISTORY = [
 */
 const Mitologia_ = () => {
 
-    const history =  useNavigate();
-
-    const handleCerrarSesion = (target,pathBack) => {
-        history(target, {  state: { pathBack } });
-    }
-
-    React.useEffect(() => {
-        document.getElementById('root').classList.add('remove_gap');
-    }, []);
-
     return (
         <>
             <Header login={true} />
-            <div className='_container_primary_profile'>
-                <Navegator />
-                <Section title='Mitologia' >
-                    {
-                        DATA_HISTORY.map((item, index) => {
-                            return (
-                                <div onClick={ () => { item.disabled && handleCerrarSesion(item.target, item.pathBack) }}  key={index} className='_container_history_section'>
-                                    <img src={require(`../../assets/item_menu/${item.img}`)} alt="icon" />
-                                    <p>{item.title}</p>
-                                </div>
-                            )
-                        })
-
-                    }
-                </Section>
-            </div>
+            <RenderSections content_section={DATA_HISTORY} title='Mitología' />
         </>
     );
